pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-reference-pipeline"
version = "0.1.0"
description = "Production-ready RAG pipeline with hybrid search and multi-provider embeddings"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3,<0.4",
    "langchain-openai>=0.2,<0.3",
    "langchain-anthropic>=0.3,<0.4",
    "langchain-chroma>=0.2,<0.3",
    "chromadb>=0.5,<0.7",
    "rank-bm25>=0.2,<0.3",
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.35",
    "python-dotenv>=1.0,<2.0",
    "pydantic>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "ruff>=0.8,<0.9",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
